requirements.txt
Python 3.6+
Pillow (PIL)
numpy